// api/types.ts
import type { IncomingMessage, ServerResponse } from 'http'

export interface VercelRequest extends IncomingMessage {
  query: { [key: string]: string | string[] }
  cookies: { [key: string]: string }
  body: any
}

export interface VercelResponse extends ServerResponse {
  send: (body: any) => VercelResponse
  json: (jsonBody: any) => VercelResponse
  status: (statusCode: number) => VercelResponse
  redirect: (statusOrUrl: string | number, url?: string) => VercelResponse
}